chore: extract getModel
