+++ lib/agents/task-manager.tsx
@@ -1,6 +1,6 @@
 import { CoreMessage, generateObject } from "ai";
-import { nextActionSchema } from "../schema/next-action";
-import { getModel } from "../utils";
+import { nextActionSchema } from "@/lib/schema/next-action";
+import { getModel } from "@/lib/utils/get-model";
 
 // Decide whether inquiry is required for the user input
 export async function taskManager(messages: CoreMessage[]) {
